backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Training program viewer - Telegram Mini App backend"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
